Break up student page class
