+++ util/spec_helper.rb
@@ -87,6 +87,7 @@ require_relative '../pages/boac/boac_class_matrix_view_page'
 require_relative '../pages/boac/boac_group_page'
 require_relative '../pages/boac/boac_filtered_cohort_page'
 require_relative '../pages/boac/boac_search_results_page'
+require_relative '../pages/boac/boac_student_page_advising_note'
 require_relative '../pages/boac/boac_student_page'
 require_relative '../pages/boac/boac_api_admin_page'
 require_relative '../pages/boac/boac_api_notes_attachment_page'
